account-statements-service/src/main/java/com/account/statement/domain/Account.java
package com.account.statement.domain;

public class Account {
	
	
	private long ID;
	private String account_type;
	private String account_number;
	
	public long getID() {
		return ID;
	}
	public void setID(long iD) {
		ID = iD;
	}
	public String getAccount_type() {
		return account_type;
	}
	public void setAccount_type(String account_type) {
		this.account_type = account_type;
	}
	public String getAccount_number() {
		return account_number;
	}
	public void setAccount_number(String account_number) {
		this.account_number = account_number;
	}
	
	

}
